requirements.txt
numpy
qiskit
qiskit-aer
tqdm
matplotlib
pylatexenc
qiskit-ibm-runtime
qiskit_qasm3_import
Pillow
torch
torchvision
h5py
polars
pandas
apache-airflow==2.11.0
apache-airflow-providers-standard
psycopg2-binary
kagglehub
huggingface_hub